Update IABot to v1.2.1
Fix a bug where cite templates’ deadurl parameter will remain untouched
while an archive is present, even if set to no.
Fix a rare bug that misinterprets comments in the URL fields of a cite
template when directly in contact with the URL.
Fix a bug that causes PHP itself to crash when stray formatting
templates are on the page.
Expand URL analysis to all links, including those with archives.