Update GUI to v1.2beta2
Fix a bug where certain filters caused the API to respond with a 404.
Fixed a bug where a misencoded response caused a blank output from a
failure to generate a JSON object.
Added a visible error to display should it happen again.
Send the browser the correct content type for the JSON response.
Pretty formatting.