T299576
Force a redirect back to the last accessed, or default set wiki when trying to access a wiki that is either closed off, or has a broken OAuth setup.
Automatically disable the wiki on detection of a broken or closed OAuth setup.
Delist the wiki from the UI for disabled wikis.
Protect active user session from being broken when attempting to access a broken wiki.