Adding msg for not posting to talk page to i18n
Sometimes, users post to talk pages that have soft redirects (to our Meta talk page) on them. Hard redirects between wikis is not possible. The soft redirect is apparently not clear enough in some instances. Therefore, we need a much more unambiguous message directing people to the right place.

For this message to be most effective, it should be in the language of the wiki it is being posted to. This patch therefore adds a "do not post to this talk page" message to our i18n project. This message will not be exposed through IABot's management interface, but this actively puts the string in front of translators so that we have a ready source of translations to go to.